+++ atd-vze/src/Components/GridTableDoughnut.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import get from "lodash.get";
-import { renderAlert } from "./gridTableHelpers";
+import { renderAlert, getChartData, getTopValues } from "./gridTableHelpers";
 
 import { Badge, Card, CardHeader, CardBody } from "reactstrap";
 import palette from "google-palette";
@@ -13,44 +13,19 @@ const GridTableDoughnut = ({ chartData, chartConfig }) => {
   const [doughnutLabels, setDoughnutLabels] = useState([]);
 
   useEffect(() => {
-    const getTopValues = (data, config) => {
-      let valueAndLabelArray = [];
-      data.forEach((record, i) => {
-        valueAndLabelArray.push([record, config.labels[i]]);
-      });
-      valueAndLabelArray = valueAndLabelArray
-        .sort((a, b) => b[0] - a[0])
-        .slice(0, config.limit);
-      const topValues = valueAndLabelArray.map((arr, i) => arr[0]);
-      const topLabels = valueAndLabelArray.map((arr, i) => arr[1]);
+    if (Object.keys(chartData).length > 0) {
+      const horizontalData = getChartData(chartData, chartConfig.doughnutChart);
+      const topChartData = getTopValues(
+        horizontalData,
+        chartConfig.doughnutChart
+      );
+
+      const topValues = topChartData.map(arr => arr[0]);
+      const topLabels = topChartData.map(arr => arr[1]);
       setDoughnutData(topValues);
       setDoughnutLabels(topLabels);
-    };
-
-    const getChartData = config => {
-      const data = config.labels.map(type => {
-        let typeTotal = 0;
-        chartData[config.table].forEach(record => {
-          if (config.isSingleRecord) {
-            if (record[config.nestedKey][config.nestedPath] === type) {
-              typeTotal++;
-            }
-          } else {
-            record[config.nestedKey].forEach(unit => {
-              if (get(unit, config.nestedPath) === type) {
-                typeTotal++;
-              }
-            });
-          }
-        });
-        return typeTotal;
-      });
-      return getTopValues(data, config);
-    };
-
-    Object.keys(chartData).length > 0 &&
-      getChartData(chartConfig.doughnutChart);
-  }, [chartData]);
+    }
+  }, [chartData, chartConfig.doughnutChart]);
 
   const graphConfig = {
     palette: palette("mpn65", chartConfig.doughnutChart.labels.length).map(
